requirements.txt
Flask~=2.1.2
Pillow~=9.1.1
numpy~=1.22.3